remove ClientHostBuilder
